requirements.txt
-e .
pytest
aws-cdk.aws-lambda
